chore(query-bar): render query bar component only when query bar store rendered
